refactor: add elastic layer
