+++ app/elastic/geo_search.py
@@ -1,7 +1,7 @@
 from elasticsearch_dsl import Q
 
-from app.search.filters.term_filters import filter_term_list_search_unite_legale
-from app.search.helpers.exclude_etablissements import (
+from app.elastic.filters.term_filters import filter_term_list_search_unite_legale
+from app.elastic.helpers.exclude_etablissements import (
     exclude_etablissements_from_search,
 )
