+++ app/elastic/helpers/bilan_filters_used.py
