+++ app/elastic/helpers/elastic_fields.py
