+++ app/elastic/helpers/exclude_etablissements.py
@@ -1,4 +1,4 @@
-def exclude_etablissements_from_search(es_search_builder):  #
+def exclude_etablissements_from_search(es_search_builder):
     # By default, exclude etablissements list from response
     include_etablissements = (
         es_search_builder.search_params.include_admin
