+++ app/elastic/helpers/helpers.py
@@ -1,4 +1,4 @@
-from app.search.parsers.siren import is_siren
+from app.elastic.parsers.siren import is_siren
 
 
 def extract_ul_and_etab_from_es_response(structure):
