+++ app/elastic/queries/text.py
